Move CallableAction and BuiltinExecutor into action module
