+++ contextshell/backends/Filesystem.py
@@ -2,8 +2,7 @@ from pathlib import Path
 from typing import List
 
 from ..NodePath import NodePath
-from ..BuiltinExecutor import BuiltinExecutor
-from ..CallableAction import action_from_function
+from ..action import action_from_function, BuiltinExecutor
 from ..exceptions import NotSupportedError
