+++ tests/functional/FilesystemTests.py
@@ -1,7 +1,7 @@
 import os
 import tempfile
 from pathlib import Path
-from contextshell.ActionExecutor import ActionExecutor
+from contextshell.action import ActionExecutor
 from tests.functional.ShellTestsBase import TreeRootTestsBase
 from tests.functional.TestExecutor import script_test
 from contextshell.backends.Filesystem import FilesystemRoot
