+++ tests/unit/BuiltinExecutorTests.py
@@ -2,7 +2,7 @@ import unittest
 
 from tests.unit.Fakes import FakeAction
 from contextshell.NodePath import NodePath
-from contextshell.BuiltinExecutor import BuiltinExecutor
+from contextshell.action import BuiltinExecutor
 
 
 class RegisterAction(unittest.TestCase):
